Отрисовыввет элементы по мокам
